+++ events/messages/responses/fortuneTeller.js
