+++ events/messages/responses/takeALook.js
